Refactor (déplacement de composants), utiliser le composant générique TitrePage sur toutes les vues de l'app
