+++ client/src/components/ComposantsGénériques/Header/index.tsx
@@ -6,7 +6,7 @@ function Header() {
     return (
         <View style={styles.header_container}>
             <Text style={styles.header_title}>Pot d'Yaourt</Text>
-            <Image style={styles.profile_image} source={require('../../assets/default-photo.png')} />
+            <Image style={styles.profile_image} source={require('../../../assets/default-photo.png')} />
         </View>
     );
 }
