+++ client/src/components/ComposantsGénériques/TitrePage/index.tsx
@@ -2,13 +2,15 @@ import React from 'react';
 import {FlatList, Image, Text, View} from 'react-native';
 import styles from "./styles";
 
+// Rendre le composant générique pour toutes les pages
+
                 //'../../assets/iconTicket.png'
 function TitrePage({titre, imageSrc}: any) {
     console.log(imageSrc);
     let imageComponent;
     switch (imageSrc) {
         case 'GardeManger':
-            imageComponent = <Image style={styles.image} source={require('../../assets/GardeMangerIcon.jpg')} />;
+            imageComponent = <Image style={styles.image} source={require('../../../assets/GardeMangerIcon.jpg')} />;
             break;
     }
     return (
