+++ client/src/screens/ListeTicket/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import { ListeTicketScreenProp } from '../../navigator'
-import Header from "../../components/Header";
+import Header from "../../components/ComposantsGénériques/Header";
 import TicketListe from "../../components/ListeTicket";
 import ListeTicketTitre from "../../components/ListeTicketTitre";
